requirements.txt
fastapi
uvicorn
requests
pandas
numpy
scikit-learn
matplotlib
python-dotenv
docker
transformers
torch
